Added MolTokenizer
